Refactor code so that new scene can be added
