+++ src/main/java/gui/stylesheet.css
@@ -1,6 +1,6 @@
 .root{
     -fx-font-size: 16pt;
-    -fx-font-family: "Courier New";
+    -fx-font-family: "Verdana";
     -fx-base: rgb(132, 145, 47);
-    -fx-background-image: url("././images/purple.jpg");
+    -fx-background-image: url("././images/light.jpg");
 }
